v3

+++ client/src/bricks/Items.js
@@ -1,51 +1,95 @@
-import React, { useState } from "react";
-import { Button, Modal, Form, Table } from 'react-bootstrap'
+import React, { useState, useEffect } from "react";
+import { Button, Modal, Form, Table, Spinner } from 'react-bootstrap'
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Icon from '@mdi/react';
 import { mdiPlus } from '@mdi/js';
+import Customspinner from "./Spinner"
 
 function Items(props) {
-  const [showModal, setModalShow] = useState(false); 
-  const [newItemName, setNewItemName] = useState("");
 
-  const closeModal = () => setModalShow(false);
-  const openModal = () => setModalShow(true);
-
-  const itemFilter = () => {
-    setShowOnlyRequired(!showOnlyRequired);
-  };
-  const [showOnlyRequired, setShowOnlyRequired] = useState(false);      
-    
-  const setItemAsPurchased = (itemId) => {
-    const updatedItems = filteredItems.map(item => {
-      if (item.id === itemId) {
-        return { ...item, required: false };
+  const [loading, setLoading] = useState(true)
+  const [error, setError] = useState(null)
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        await new Promise((resolve) => {setTimeout(() => {resolve();}, 1000)})
+        setLoading(false)
+      } catch (error) {
+        setError(error)
+        setLoading(false)
       }
-      return item;
-    });
-    setItems(updatedItems);
-  };
+    }
+    fetchData()
+  }, [])
 
+
+  const [showModal, setModalShow] = useState(false)
+  const [newItemName, setNewItemName] = useState("")
+
+  const closeModal = () => setModalShow(false)
+  const openModal = () => setModalShow(true)
+
+  const itemFilter = () => {setShowOnlyRequired(!showOnlyRequired)}
+  const [showOnlyRequired, setShowOnlyRequired] = useState(false)
   const [items, setItems] = useState(props.items)
           
-  const filteredItems = showOnlyRequired ? items.filter(item => item.required) : items;
+  const filteredItems = showOnlyRequired ? items.filter(item => item.required) : items;     
 
-  const deleteItem = (itemId) => {
-    const updatedItems = items.filter(item => item.id !== itemId);
-    setItems(updatedItems);
-  };
+  const fakeServerRequest = () => {
+    return new Promise((resolve) => {setTimeout(() => {resolve("OK")}, 1000)})}
 
-  const addItem = () => {
-    if (newItemName.trim() === '') {return}
-      const newItem = { id: Math.random().toString(), name: newItemName, required: true,};
-      const updatedItems = [...items, newItem]; 
-      setItems(updatedItems); 
-      closeModal(); 
-      setNewItemName(""); 
-  };
+//Označování položky jako zakoupenou
+  const [purchasingItem, setPurchasingItem] = useState(false)
+  const [purchasingItemId, setPurchasingItemId] = useState(null)
+  const setItemAsPurchased = async (itemId) => {
+    try {
+      setPurchasingItem(true)
+      await fakeServerRequest()
+      const updatedItems = filteredItems.map(item => {if (item.id === itemId) {return { ...item, required: false }}return item})
+      setItems(updatedItems)
+      setPurchasingItem(false)
+    } catch (error) {
+      console.error('Chyba při označování položky za zakoupenou:', error)
+    } 
+  }
+
+//Mazání položky
+  const [deletingItem, setDeletingItem] = useState(false)
+  const [deletingItemId, setDeletingItemId] = useState(null)
+  const deleteItem = async (itemId) => {
+    try {
+        setDeletingItem(true)
+        await fakeServerRequest()
+        const updatedItems = items.filter(item => item.id !== itemId)
+        setItems(updatedItems)
+        setDeletingItem(true)
+    } catch (error) {
+        console.error('Chyba při mazání položky:', error)
+    } 
+}
+
+//Přidávání položky
+  const [addingItem, setAddingItem] = useState(false)
+  const addItem = async () => {
+    try {
+        setAddingItem(true)
+        await fakeServerRequest()
+        if (newItemName.trim() === '') {return}
+        const newItem = { id: Math.random().toString(), name: newItemName, required: true}
+        const updatedItems = [...items, newItem]
+        setItems(updatedItems)
+        setAddingItem(false)
+        closeModal() 
+        setNewItemName("")
+    } catch (error) {
+        console.error('Chyba při přidávání položky:', error)
+    } 
+}
         
   return (
     <div style={{padding: "10px"}} class="csscontainer">
+      {error ? (<p>Chyba při získávání dat</p>):(<>
+      {loading ? (<Customspinner variant="Bounce"/>):(
       <Table hover>
         <thead>
           <tr>
@@ -61,15 +105,15 @@ function Items(props) {
               <td style={{ textDecoration: item.required ? ("none") : ("line-through") }}>{item.name}</td>
               <td style={{ textAlign: 'right' }}>
                 {item.required ? (
-                  <Button variant="outline-primary" onClick={() => setItemAsPurchased(item.id)}>Zakoupené</Button>
+                  <Button variant="outline-primary" onClick={() => {setItemAsPurchased(item.id); setPurchasingItemId(item.id)}}>{purchasingItem && purchasingItemId === item.id ? (<Spinner animation="border" role="status" size="sm"></Spinner>) : ("Zakoupené")}</Button>
                 ):(
-                  <Button variant="outline-success" disabled>Zakoupeno</Button> //Značka zakoupení
+                  <Button variant="outline-success" disabled>Zakoupeno</Button>
                 )} 
-                <Button variant="outline-danger" onClick={() => deleteItem(item.id)}>Smazat</Button></td>
+                <Button variant="outline-danger" onClick={() => {deleteItem(item.id); setDeletingItemId(item.id)}}>{deletingItem && deletingItemId === item.id ? (<Spinner animation="border" role="status" size="sm"></Spinner>) : ("Smazat")}</Button></td>
             </tr>
           ))}
         </tbody>
-      </Table>
+      </Table>)}</>)}
       <Modal show={showModal} onHide={closeModal}>
         <Modal.Header>
           <Modal.Title>Přidat novou položku</Modal.Title>
@@ -81,12 +125,13 @@ function Items(props) {
         </Modal.Body>
         <Modal.Footer>
           <Button variant="secondary" onClick={closeModal}>Zavřít</Button>
-          <Button variant="primary" onClick={addItem}>Přidat</Button>
+          <Button variant="primary" onClick={addItem} disabled={newItemName.trim() === ''}> {addingItem ? (<Spinner animation="border" role="status" size="sm"></Spinner>):("Přidat")}</Button>
         </Modal.Footer>
       </Modal>
+      {loading ? (null) : (
       <Button variant="success" onClick={openModal} style={{margin: "5px 0 10px"}}>
         <Icon path={mdiPlus} size={1} /> Přidat novou položku
-      </Button>
+      </Button>)}
     </div>
   );
 }
